Made Foliage Placers more consistent
