+++ src/main/java/com/aetherteam/aetherii/world/tree/foliage/GreatoakFoliagePlacer.java
@@ -22,11 +22,6 @@ public class GreatoakFoliagePlacer extends FoliagePlacer {
         super(radius, offset);
     }
 
-    @Override
-    protected FoliagePlacerType<?> type() {
-        return AetherIIFoliagePlacerTypes.GREATOAK_FOLIAGE_PLACER.get();
-    }
-
     /**
      * Places a sphere of leaves.
      *
@@ -82,4 +77,10 @@ public class GreatoakFoliagePlacer extends FoliagePlacer {
     protected boolean shouldSkipLocation(RandomSource random, int localX, int localY, int localZ, int range, boolean large) {
         return Mth.square(localX) + Mth.square(localY - 1) + Mth.square(localZ) > range + random.nextInt(3);
     }
+
+    @Override
+    protected FoliagePlacerType<?> type() {
+        return AetherIIFoliagePlacerTypes.GREATOAK_FOLIAGE_PLACER.get();
+    }
+
 }
