+++ src/main/java/com/aetherteam/aetherii/world/tree/foliage/LargeSkyrootFoliagePlacer.java
@@ -22,11 +22,6 @@ public class LargeSkyrootFoliagePlacer extends FoliagePlacer {
         super(radius, offset);
     }
 
-    @Override
-    protected FoliagePlacerType<?> type() {
-        return AetherIIFoliagePlacerTypes.LARGE_SKYROOT_FOLIAGE_PLACER.get();
-    }
-
     /**
      * Places a sphere of leaves.
      *
@@ -97,5 +92,10 @@ public class LargeSkyrootFoliagePlacer extends FoliagePlacer {
     @Override
     protected boolean shouldSkipLocation(RandomSource random, int localX, int localY, int localZ, int range, boolean large) {
         return Mth.square(localX) + Mth.square(localY + 2) + Mth.square(localZ) > range + random.nextInt(3);
+    }
+
+    @Override
+    protected FoliagePlacerType<?> type() {
+        return AetherIIFoliagePlacerTypes.LARGE_SKYROOT_FOLIAGE_PLACER.get();
     }
 }
