+++ src/main/java/com/aetherteam/aetherii/world/tree/foliage/WisprootFoliagePlacer.java
@@ -20,11 +20,6 @@ public class WisprootFoliagePlacer extends DarkOakFoliagePlacer {
     }
 
     @Override
-    protected FoliagePlacerType<?> type() {
-        return AetherIIFoliagePlacerTypes.WISPROOT_FOLIAGE_PLACER.get();
-    }
-
-    @Override
     protected void createFoliage(LevelSimulatedReader level, FoliageSetter foliageSetter, RandomSource random, TreeConfiguration config, int maxFreeTreeHeight, FoliageAttachment attachment, int foliageHeight, int foliageRadius, int offset) {
         for (int i = offset; i >= offset - foliageHeight; --i) {
             this.placeLeavesRow(level, foliageSetter, random, config, attachment.pos(), 14, i, attachment.doubleTrunk());
@@ -50,5 +45,10 @@ public class WisprootFoliagePlacer extends DarkOakFoliagePlacer {
     @Override
     protected boolean shouldSkipLocation(RandomSource random, int localX, int localY, int localZ, int range, boolean large) {
         return Mth.square(localX) + Mth.square(localY - 1) + Mth.square(localZ) > range + random.nextInt(2);
+    }
+
+    @Override
+    protected FoliagePlacerType<?> type() {
+        return AetherIIFoliagePlacerTypes.WISPROOT_FOLIAGE_PLACER.get();
     }
 }
